Refactored folder structure & styled category card mouse pointer
