+++ src/components/App.js
@@ -41,11 +41,6 @@ const useStyles = makeStyles(() => ({
     "*::-webkit-scrollbar": {
       width: "0.4em",
     },
-    // "*::-webkit-scrollbar-track": {
-    //   background: "rgb(0, 0, 0, 0.12)",
-    //   border: "4px solid transparent",
-    //   backgroundClip: "content-box",
-    // },
     "*::-webkit-scrollbar-thumb": {
       backgroundColor: "#acacac",
       borderRadius: "50px",
