+++ src/components/Item/AddPosting/AddPosting.js
@@ -5,10 +5,10 @@ import Button from "@material-ui/core/Button";
 import Paper from "@material-ui/core/Paper";
 import Typography from "@material-ui/core/Typography";
 import { connect } from "react-redux";
-import { addPosting, loadItemDetail } from "../../redux/actions/postingActions";
-import { closeModal } from "../../redux/actions/modalActions";
-import { displayError } from "../../redux/actions/snackbarActions";
-import { ADD_POSTING_ERROR } from "../../redux/constants/snackbarMessageTypes";
+import { addPosting, loadItemDetail } from "../../../redux/actions/postingActions";
+import { closeModal } from "../../../redux/actions/modalActions";
+import { displayError } from "../../../redux/actions/snackbarActions";
+import { ADD_POSTING_ERROR } from "../../../redux/constants/snackbarMessageTypes";
 
 import DisplayStepper from "./DisplayStepper";
 import Step1 from "./Step1";
