+++ src/components/Item/SearchBar.js
@@ -15,7 +15,7 @@ import {
 import Select from "@material-ui/core/Select";
 import MenuItem from "@material-ui/core/MenuItem";
 import FormControl from "@material-ui/core/FormControl";
-import { categories, conditions } from "../../redux/constants/classifierTypes";
+import { categories, conditions } from "../constants/classifierTypes";
 import Button from "@material-ui/core/Button";
 import FormHelperText from "@material-ui/core/FormHelperText";
 import clsx from "clsx";
