+++ src/components/Offering/OffersSent/OffersSent.js
@@ -4,7 +4,7 @@ import { Grid } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
 import { OfferSentPreviewCard } from "./OfferSentPreviewCard";
-import ConfirmationDialog from "../../ConfirmationDialog";
+import ConfirmationDialog from "../../shared/ConfirmationDialog";
 import { rescindOffer } from "../../../redux/actions/offeringActions";
 
 const useStyles = makeStyles(() => ({
