+++ src/components/PathRouter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 
-import ItemPage from "./Item/ItemPage";
+import ItemPage from "./pages/ItemPage";
 import { ItemResults } from "./pages/ItemResults";
 import LandingPage from "./pages/LandingPage";
 import UserProfile from "./pages/UserProfile";
