+++ src/components/constants/classifierTypes.js
