+++ src/components/pages/ItemPage.js
@@ -5,10 +5,10 @@ import { useHistory, useLocation } from "react-router";
 import { makeStyles } from "@material-ui/core/styles";
 import { Button, Container, Divider } from "@material-ui/core";
 
-import ReviewSection from "./ReviewSection";
-import ItemDetailContainer from "./ItemDetailContainer";
+import ReviewSection from "../Item/ReviewSection";
+import ItemDetailContainer from "../Item/ItemDetailContainer";
 import { loadItemDetail } from "../../redux/actions/postingActions";
-import MessageButton from "./MessageButton";
+import MessageButton from "../Item/MessageButton";
 
 const useStyles = makeStyles((theme) => ({
   root: {
